requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
anthropic==0.5.0
google-api-python-client==2.97.0
google-auth-oauthlib==1.1.0
requests==2.31.0
webvtt-py==0.4.6
python-multipart==0.0.6
jinja2==3.1.2
PyJWT==2.8.0
aiofiles==23.2.1
pandas==2.0.3
